+++ Breaks6502/BreaksDebug/FormAbout.cs
@@ -1,14 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace BreaksDebug
+﻿namespace BreaksDebug
 {
 	public partial class FormAbout : Form
 	{
