+++ Breaks6502/BreaksDebug/FormMain.cs
@@ -1,14 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-using System.Runtime.InteropServices;
+﻿using System.Runtime.InteropServices;
 using Be.Windows.Forms;
 using System.IO;
 using Newtonsoft.Json;
@@ -22,58 +12,11 @@ namespace BreaksDebug
 		static extern bool AllocConsole();
 
 		CoreDebug sys = new CoreDebug();
-		byte[] mem = new byte[0x10000];
-		IByteProvider memProvider;
 		string testAsmName = "Test.asm";
 		string MarkdownDir = "WikiMarkdown";        // To dump the state as Markdown text, you need to create this directory.
 		string MarkdownImgDir = "imgstore/ops";
 		string WikiRoot = "/BreakingNESWiki/";
 		bool MarkdownOutput = false;
-
-		long Cycle = 0;
-
-		class UnitTestParam
-		{
-			public bool CompileFromSource = true;
-			public string MemDumpInput = "mem.bin";
-			public string AsmSource = "Test.asm";
-			public string RamStart = "0";
-			public string RamSize = "0x800";
-			public string RomStart = "0xc000";
-			public string RomSize = "0x4000";
-			public bool RunUntilBrk = true;
-			public bool RunCycleAmount = true;
-			public long CycleMax = 10000;
-			public bool RunUntilPC = false;
-			public string PC = "0x0";
-			public bool TraceMemOps = true;
-			public bool TraceCLK = true;
-			public bool DumpMem = true;
-			public string JsonResult = "res.json";
-			public string MemDumpOutput = "mem2.bin";
-		}
-
-		class UnitTestResult
-		{
-			public string A { get; set; } = "0x00";
-			public string X { get; set; } = "0x00";
-			public string Y { get; set; } = "0x00";
-			public string S { get; set; } = "0x00";
-			public string PC { get; set; } = "0x0000";
-			public int C { get; set; } = 0;
-			public int Z { get; set; } = 0;
-			public int I { get; set; } = 0;
-			public int D { get; set; } = 0;
-			public int V { get; set; } = 0;
-			public int N { get; set; } = 0;
-			public UInt64 CLK { get; set; } = 0;
-		}
-
-		bool UnitTestMode = false;
-		string TestInputJson;
-		UnitTestParam testParam;
-		UInt64 UnitTestPhiCounter = 0;
-		UInt64 UnitTestPhiTraceCounter = 0;
 
 		public FormMain()
 		{
@@ -86,6 +29,8 @@ namespace BreaksDebug
 			AllocConsole();
 #endif
 
+			BreaksCoreInterop.CreateBoard("Bogus", "None", "None", "None");
+
 			SetStyle(ControlStyles.OptimizedDoubleBuffer, true);
 
 			toolStripButton4.Checked = true;    // /NMI
@@ -95,9 +40,6 @@ namespace BreaksDebug
 			toolStripButton2.Checked = true;    // RES
 			ButtonsToPads();
 
-			memProvider = new DynamicByteProvider(mem);
-			hexBox1.ByteProvider = memProvider;
-			//sys.AttatchMemory(memProvider);
 			UpdateAll();
 
 			// Select the operating mode - manual or automatic unit test.
@@ -135,138 +77,6 @@ namespace BreaksDebug
 			}
 		}
 
-		int Strtol(string str)
-		{
-			str = str.Trim();
-
-			return (str.Contains("0x") || str.Contains("0X")) ?
-				Convert.ToInt32(str, 16) :
-					str[0] == '0' ? Convert.ToInt32(str, 8) : Convert.ToInt32(str, 10);
-		}
-
-		int InitUnitTest()
-		{
-			BogusSystem.MemoryMapping map = new BogusSystem.MemoryMapping();
-
-			map.RamStart = Strtol(testParam.RamStart);
-			map.RamSize = Strtol(testParam.RamSize);
-			map.RomStart = Strtol(testParam.RomStart);
-			map.RomSize = Strtol(testParam.RomSize);
-
-			//sys.SetMemoryMapping(map);
-
-			if (testParam.CompileFromSource)
-			{
-				LoadAsm(testParam.AsmSource);
-				Assemble();
-			}
-			else
-			{
-				LoadMemDump(testParam.MemDumpInput);
-			}
-
-			// Perform the PreBRK PowerUP sequence
-
-			for (int i = 0; i < 8; i++)
-			{
-				Step();
-			}
-
-			// Cancel Reset and wait for Reset BRK to complete
-
-			toolStripButton6.Checked = true;
-
-			int timeout = 0x100;
-
-			while (timeout-- != 0)
-			{
-				Step();
-
-				var regsBuses = sys.GetRegsBuses();
-				if (regsBuses.IRForDisasm != 0x00)
-				{
-					break;
-				}
-			}
-
-			if (timeout == 0)
-			{
-				Console.WriteLine("The operation of the simulator is broken. BRK-sequence is not completed.");
-				return -1;
-			}
-
-			return 0;
-		}
-
-		void StartUnitTest()
-		{
-			ushort BreakOnPC = (ushort)Strtol(testParam.PC);
-
-			while (true)
-			{
-				Step();
-				UnitTestPhiCounter++;
-				UnitTestPhiTraceCounter++;
-				if (UnitTestPhiTraceCounter > 1000000 * 2)
-				{
-					// Output statistics every million cycles
-					if (testParam.TraceCLK)
-					{
-						Console.WriteLine("CLK: " + (UnitTestPhiCounter / 2).ToString());
-					}
-					UnitTestPhiTraceCounter = 0;
-				}
-
-				if (Cycle >= testParam.CycleMax && testParam.RunCycleAmount)
-				{
-					break;
-				}
-
-				var regsBuses = sys.GetRegsBuses();
-				if (regsBuses.IRForDisasm == 0x00 && testParam.RunUntilBrk)
-				{
-					break;
-				}
-				if (regsBuses.PCForUnitTest == BreakOnPC && testParam.RunUntilPC)
-				{
-					break;
-				}
-			}
-
-			SaveUnitTestResults();
-		}
-
-		void SaveUnitTestResults()
-		{
-			if (testParam.DumpMem)
-			{
-				SaveMemDump(testParam.MemDumpOutput);
-			}
-
-			UnitTestResult res = new UnitTestResult();
-
-			var regsBuses = sys.GetRegsBuses();
-
-			res.A = regsBuses.AC;
-			res.X = regsBuses.X;
-			res.Y = regsBuses.Y;
-			res.S = regsBuses.S;
-			res.PC = "0x" + regsBuses.PCForUnitTest.ToString("X4");
-			res.C = regsBuses.C_OUT;
-			res.Z = regsBuses.Z_OUT;
-			res.I = regsBuses.I_OUT;
-			res.D = regsBuses.D_OUT;
-			res.V = regsBuses.V_OUT;
-			res.N = regsBuses.N_OUT;
-			res.CLK = UnitTestPhiCounter / 2;
-
-			string json = JsonConvert.SerializeObject(res);
-
-			File.WriteAllText(testParam.JsonResult, json);
-
-			Close();
-		}
-
 		private void exitToolStripMenuItem_Click(object sender, EventArgs e)
 		{
 			Close();
@@ -297,7 +107,8 @@ namespace BreaksDebug
 				trace = testParam.TraceMemOps;
 			}
 
-			//sys.Step(trace);
+			BreaksCoreInterop.Step();
+
 			if (!UnitTestMode)
 			{
 				UpdateAll();
@@ -306,7 +117,7 @@ namespace BreaksDebug
 
 		void UpdateCycleStats()
 		{
-			toolStripStatusLabel1.Text = "Cycle: " + Cycle.ToString();
+			toolStripStatusLabel1.Text = "Cycle: " + BreaksCoreInterop.GetPHICounter().ToString();
 		}
 
 		void UpdateMemoryDump()
@@ -376,30 +187,55 @@ namespace BreaksDebug
 			UpdateCpuDebugInfo();
 		}
 
-		void LoadMemDump(string file)
+		void LoadMemDump(byte[] dump)
 		{
-			byte[] dump = File.ReadAllBytes(file);
+			List<BreaksCore.MemDesciptor> mem = BreaksCore.GetMemoryLayout();
 
-			for (int i = 0; i < mem.Length; i++)
+			int descrID = -1;
+
+			for (int i = 0; i < mem.Count; i++)
 			{
-				memProvider.WriteByte(i, dump[i]);
+				if (mem[i].name == "WRAM")
+				{
+					descrID = i;
+					break;
+				}
 			}
 
-			hexBox1.Refresh();
-			Console.WriteLine("Loaded memory dump: " + file);
+			if (descrID >= 0)
+			{
+				BreaksCore.WriteMem(descrID, dump);
+				if (!UnitTestMode)
+				{
+					hexBox1.ByteProvider = new DynamicByteProvider(dump);
+					hexBox1.Refresh();
+				}
+			}
 		}
 
-		void SaveMemDump(string file)
+		byte[] SaveMemDump()
 		{
-			byte[] dump = new byte[mem.Length];
+			List<BreaksCore.MemDesciptor> mem = BreaksCore.GetMemoryLayout();
 
-			for (int i = 0; i < mem.Length; i++)
+			int descrID = -1;
+
+			for (int i = 0; i < mem.Count; i++)
 			{
-				dump[i] = memProvider.ReadByte(i);
+				if (mem[i].name == "WRAM")
+				{
+					descrID = i;
+					break;
+				}
 			}
 
-			File.WriteAllBytes(file, dump);
-			Console.WriteLine("Saved memory dump: " + file);
+			if (descrID >= 0)
+			{
+				byte[] buf = new byte[mem[descrID].size];
+				BreaksCore.DumpMem(descrID, buf);
+				return buf;
+			}
+
+			return null;
 		}
 
 		private void loadMemoryDumpToolStripMenuItem_Click(object sender, EventArgs e)
@@ -407,7 +243,9 @@ namespace BreaksDebug
 			DialogResult res = openFileDialog2.ShowDialog();
 			if (res == DialogResult.OK)
 			{
-				LoadMemDump(openFileDialog2.FileName);
+				byte[] dump = File.ReadAllBytes(openFileDialog2.FileName);
+				LoadMemDump(dump);
+				Console.WriteLine("Loaded memory dump: " + openFileDialog2.FileName);
 			}
 		}
 
@@ -416,7 +254,9 @@ namespace BreaksDebug
 			DialogResult res = saveFileDialog1.ShowDialog();
 			if (res == DialogResult.OK)
 			{
-				SaveMemDump(saveFileDialog1.FileName);
+				var dump = SaveMemDump();
+				File.WriteAllBytes(saveFileDialog1.FileName, dump);
+				Console.WriteLine("Saved memory dump: " + saveFileDialog1.FileName);
 			}
 		}
 
@@ -445,7 +285,7 @@ namespace BreaksDebug
 		{
 			Console.WriteLine("Assemble");
 
-			byte[] buffer = new byte[mem.Length];
+			byte[] buffer = new byte[0x10000];
 
 			int num_err = BreaksCoreInterop.Assemble(richTextBox1.Text, buffer);
 			if (num_err != 0)
@@ -456,12 +296,7 @@ namespace BreaksDebug
 				return;
 			}
 
-			for (int i = 0; i < mem.Length; i++)
-			{
-				memProvider.WriteByte(i, buffer[i]);
-			}
-
-			hexBox1.Refresh();
+			LoadMemDump(buffer);
 		}
 
 		void ButtonsToPads()
